+++ package.json
@@ -10,6 +10,7 @@
   "devDependencies": {
     "@playwright/test": "^1.41.1",
     "@types/node": "^20.11.5",
-    "dotenv": "^16.4.1"
+    "dotenv": "^16.4.1",
+    "typescript": "^5.3.3"
   }
 }
